+++ traits/NotificationLog.php
@@ -5,20 +5,19 @@
  * | | / // // ___//_  _//   ||  __||_   _|
  * | |/ // /(__  )  / / / /| || |     | |
  * |___//_//____/  /_/ /_/ |_||_|     |_|
- * @link http://vistart.name/
+ * @link https://vistart.name/
  * @copyright Copyright (c) 2016 vistart
- * @license http://vistart.name/license/
+ * @license https://vistart.name/license/
  */
 
-namespace vistart\Models\components;
+namespace vistart\Models\traits;
 
 /**
- * Description of MultiDomainsUrlManager
+ * Description of NotificationLog
  *
- * @since 2.0
  * @author vistart <i@vistart.name>
  */
-class MultiDomainsUrlManager extends \yii\web\UrlManager
+trait NotificationLog
 {
-    //put your code here
+    
 }